+++ spec/transpec/syntax/should_receive_spec.rb
@@ -405,6 +405,40 @@ module Transpec
             record.converted_syntax.should == 'expect_any_instance_of(SomeClass).to receive(:message)'
           end
         end
+
+        context 'when it is `described_class.any_instance.should_receive(:method)` form' do
+          let(:source) do
+            <<-END
+              describe 'example' do
+                it 'receives #foo' do
+                  described_class.any_instance.should_receive(:foo)
+                end
+              end
+            END
+          end
+
+          let(:expected_source) do
+            <<-END
+              describe 'example' do
+                it 'receives #foo' do
+                  expect_any_instance_of(described_class).to receive(:foo)
+                end
+              end
+            END
+          end
+
+          it 'converts into `expect_any_instance_of(described_class).to receive(:method)` form' do
+            should_receive_object.expectize!
+            rewritten_source.should == expected_source
+          end
+
+          it 'adds record "`SomeClass.any_instance.should_receive(:message)` ' +
+             '-> `expect_any_instance_of(SomeClass).to receive(:message)`"' do
+            should_receive_object.expectize!
+            record.original_syntax.should  == 'SomeClass.any_instance.should_receive(:message)'
+            record.converted_syntax.should == 'expect_any_instance_of(SomeClass).to receive(:message)'
+          end
+        end
       end
 
       describe '#useless_expectation?' do
@@ -605,7 +639,41 @@ module Transpec
             END
           end
 
-          it 'converts into `allow_any_instance_of(subject).to receive(:method)` form' do
+          it 'converts into `allow_any_instance_of(SomeClass).to receive(:method)` form' do
+            should_receive_object.allowize_useless_expectation!
+            rewritten_source.should == expected_source
+          end
+
+          it 'adds record "`SomeClass.any_instance.should_receive(:message).any_number_of_times` ' +
+             '-> `allow_any_instance_of(SomeClass).to receive(:message)`"' do
+            should_receive_object.allowize_useless_expectation!
+            record.original_syntax.should  == 'SomeClass.any_instance.should_receive(:message).any_number_of_times'
+            record.converted_syntax.should == 'allow_any_instance_of(SomeClass).to receive(:message)'
+          end
+        end
+
+        context 'when it is `described_class.any_instance.should_receive(:method).any_number_of_times` form' do
+          let(:source) do
+            <<-END
+              describe 'example' do
+                it 'responds to #foo' do
+                  described_class.any_instance.should_receive(:foo).any_number_of_times
+                end
+              end
+            END
+          end
+
+          let(:expected_source) do
+            <<-END
+              describe 'example' do
+                it 'responds to #foo' do
+                  allow_any_instance_of(described_class).to receive(:foo)
+                end
+              end
+            END
+          end
+
+          it 'converts into `allow_any_instance_of(described_class).to receive(:method)` form' do
             should_receive_object.allowize_useless_expectation!
             rewritten_source.should == expected_source
           end
@@ -673,7 +741,7 @@ module Transpec
             END
           end
 
-          it 'converts into `allow_any_instance_of(subject).to receive(:method)` form' do
+          it 'converts into `allow_any_instance_of(SomeClass).to receive(:method)` form' do
             should_receive_object.allowize_useless_expectation!
             rewritten_source.should == expected_source
           end